+++ src/components/common/StyledAlert.js
